Migrate some common methods
